+++ NEWS.md
@@ -1,8 +1,12 @@
 # r2resize 1.6
 
-## Planned Features
+## Features
 
-  - [ ] More robust integration for Shiny
+  - [x] elastiCard function for elastic image containers
+  
+  - [x] flexCard function for clickable image blocks
+  
+  - [x] splitCard2 function is another version of the splitCard
   
 # r2resize 1.5